mainをdevelopと同期させる

+++ models/wage.py
@@ -1,10 +1,10 @@
-from peewee import Model,CharField,FloatField
+from peewee import Model,CharField,IntegerField
 from .db import db
 
 class Wage(Model):
     location = CharField() # 勤務地
-    weekday_wage = FloatField() # 平日時給
-    holiday_wage = FloatField() # 休日時給
+    weekday_wage = IntegerField() # 平日時給
+    holiday_wage = IntegerField() # 休日時給
     
     class Meta:
         database = db
